+++ src/main/java/net/onlinelibrary/rest/UserController.java
@@ -1,6 +1,8 @@
 package net.onlinelibrary.rest;
 
-import net.onlinelibrary.dto.*;
+import net.onlinelibrary.authorization.UserPrincipalImpl;
+import net.onlinelibrary.dto.CommentDto;
+import net.onlinelibrary.dto.UserDto;
 import net.onlinelibrary.exception.*;
 import net.onlinelibrary.mapper.CommentMapper;
 import net.onlinelibrary.mapper.UserMapper;
@@ -8,8 +10,12 @@ import net.onlinelibrary.model.Role;
 import net.onlinelibrary.model.User;
 import net.onlinelibrary.service.UserService;
 import org.springframework.beans.BeanUtils;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpSession;
+import java.util.Collections;
 import java.util.Date;
 import java.util.Set;
 import java.util.stream.Stream;
@@ -18,6 +24,7 @@ import java.util.stream.Stream;
 @RequestMapping("api/users")
 public class UserController {
     private final UserService userService;
+
     private final UserMapper userMapper;
     private final CommentMapper commentMapper;
 
@@ -76,6 +83,7 @@ public class UserController {
         user.setLastModifiedDate(new Date());
 
         user.setActive(true);
+        user.setRoles(Collections.singleton(Role.USER));
 
         try {
             return userMapper.toDto(userService.saveUser(user));
@@ -86,9 +94,17 @@ public class UserController {
         }
     }
 
+    @PreAuthorize("hasAuthority('USER')")
     @PutMapping("{id}")
     public UserDto fullUpdateUser(@PathVariable("id") Long userId, @RequestBody UserDto dto) {
         try {
+            User authorizedUser = ((UserPrincipalImpl)SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUser();
+            if (!authorizedUser.getId().equals(userId) && !authorizedUser.getRoles().contains(Role.KOSTYAN))
+                throw new ForbiddenException("You do not have access to modify user");
+
+            if (!authorizedUser.getRoles().contains(Role.KOSTYAN) && dto.getRoles() != null)
+                throw new ForbiddenException("You do not have access to set user roles");
+
             User user = userService.getById(userId);
             User userByDto = userMapper.toEntity(dto);
 
@@ -110,9 +126,17 @@ public class UserController {
         }
     }
 
+    @PreAuthorize("hasAuthority('USER')")
     @PatchMapping("{id}")
     public UserDto partUpdateUser(@PathVariable("id") Long userId, @RequestBody UserDto dto) {
         try {
+            User authorizedUser = ((UserPrincipalImpl)SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUser();
+            if (!authorizedUser.getId().equals(userId) && !authorizedUser.getRoles().contains(Role.KOSTYAN))
+                throw new ForbiddenException("You do not have access to modify user");
+
+            if (!authorizedUser.getRoles().contains(Role.KOSTYAN) && dto.getRoles() != null)
+                throw new ForbiddenException("You do not have access to set user roles");
+
             User user = userService.getById(userId);
             User userByDto = userMapper.toEntity(dto);
 
@@ -140,11 +164,16 @@ public class UserController {
     }
 
 
-
+    @PreAuthorize("hasAuthority('USER')")
     @DeleteMapping("{id}")
-    public void deleteUser(@PathVariable("id") Long userId) {
+    public void deleteUser(@PathVariable("id") Long userId, HttpSession session) {
         try {
+            User authorizedUser = ((UserPrincipalImpl)SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUser();
+            if (!authorizedUser.getId().equals(userId) && !authorizedUser.getRoles().contains(Role.KOSTYAN))
+                throw new ForbiddenException("You do not have access to delete user");
+
             userService.deleteById(userId);
+            session.invalidate();
         } catch (UserNotFoundException e) {
             throw new NotFoundException(e.getMessage());
         }
